monterrei_core/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-socketio==5.11.4
python-multipart==0.0.12
jinja2==3.1.4
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
mido==1.3.2
python-rtmidi==1.5.8
pyserial==3.5
loguru==0.7.2
itsdangerous==2.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
